Decompose the Listing god file into focused modules
Listing.tsx had grown to 2580 lines holding the directory fetch, streamed
search walk, selection model, preview pane, file operations, context menus,
and keyboard handling in one component. Split it into 12 modules under
apps/explorer/listing/ (pure helpers + focused hooks), leaving Listing.tsx
as an 836-line orchestrator that wires the hooks and renders the table.

Behavior is unchanged: code moved verbatim except the searchSort comparator
extracted as sortHits() and a local rename to avoid shadowing.
